Setup HTML and CSS
Setup the index.html
and index.css as
instructed by the
assignment.
Add info spans inside the
divs of all input types.
These spans will inform
users of validation status.

Set the form to
a novalidate attribute
and not add HTML
validation features.

The plan is to
use JavaScript for
the form validation.