Validate Email field
Setup a class that would
validate the form. In the
class constructor, select
a form and run the
initialize method if a form
exists.

Initialize a validateEmail
method that will validate
the email field. In the
method, select the email field.
Define a regular expression
and save it as a constant.
The regular expression
allows 4 characters -
uppercas, lowercase and
numbers, then @, then
thecompany.com.

Add a focusout event listener
to the email input so that
validation happens when the
focus leaves the input.

In the event listener, check
if the length of the value
of the email input is 0
(means it is empty) and
check if the value of the
input matches the regular
expression. Save the
results of these 2 checks
in respective constants.

Set a customValidity message
if the email field is empty
and another if it does not
match the regular expression
pattern, else, set the
customValidity message to
be an empty string
(means the constraint is satisfied).

Report the validity so that the
custom validity message is
displayed without clicking
a submit button.

Call the validateEmail method
in initialize() so that
email validation is setup
when a new instance of
the class is created.