Style validations
Select all inputs that
are valid (using
pseudo class) and style
it appropriately. Hide
the outline else the
outline will stay above
the border and the
border wont be visible
unless it has a
much larger
px value.

Style the inputs
in which their placeholders
are shown, with the same
border as was given
input (line 43) so that all
inputs are not styled as
valid when the page loads.
(They are actually styled
as :valid first but due to
the fact that the style rules
of :placeholder-shown comes
below the style rules of
valid CSS cascading order
makes :placeholder-shown
style rules to override
:valid style rules)

Define style rules for
input:invalid also.