Create country and zip validation condition
First select the country and
zipcode fields in the constructor.
Create a constraints object that
contains zip code regular
expressions pattern and
validity messages for each
corresponding country
(denoted in the object by the
country's corresponding
value in the select field.)

Create a currentConstraint
property that creates a
Regular Expression out of
the pattern in the objects'
arrays first element. This is
capture Switzerland because
it is the first option in
the country select field.

Initialize a method,
validateCountryAndZipCondition().
The method first saves the
result of the test of the
value of the zipcode field
against the currentConstraint
pattern to a constant,
matchesPattern.
If it matches the pattern,
it sets the custom validity message
to an empty string else it sets the
validity message to the second element
of the array of the corresponding
country.

It then returns matchesPattern.

This method will be called when
the country select field changes
and when the zipcode field loses
focus so that country and
zipcode can be validated.