Validate zipcode and country
Initialize a validateCountry()
method that adds a change
event listener to the
country field and sets the
currentConstraint to the
event.target.value - the value
of the option that was selected.

It then calls the
validateCountryAndZipCondition().

Initialize a validateZipCode()
method that sets a focusout
event listener on the zipcode
field input and reports
the validity of the zipcode
if validateCountryAndZipCondition()
is false. If the zipcode
that was typed in did not match
the pattern that was specified for
the country, it reports
the validity of the zipcode
field, when the field loses
focus.

Call the validateCountry()
and validateZipCode in the
initialize() method.