Validate password fields
In the constructor, select the
password and confirmPassword
fields.

Initialize a validatePasswordField
method. The method first defines
a constraint - an array that
holds the RegExp pattern for the
passwords and a description of
the requirements of the pattern.
The method then creates a new
RexExp from the pattern. It
then checks if the password
field is empty and if
its value matches the RegExp
pattern - whle setting
appropriate validity
messages. If it passes all these
tests, it sets the validity
message to an empty string.
The method then reports the
validity.

Initialize a
validateConfirmPasswordField
method that first checks if
the password field is not
empty and the confirm password
is not. It also checks if the
value of the password field
matches with the value of
the confirmPasswordField -
while setting the
appropriate validation messages.
If it passes all these
tests, it sets the validity
message to an empty string.
The method then reports the
validity.

These two methods will be used to
check the validity of these
fields when the fields lose focus.