Initialize password validation
Create a method that adds focusout
event listeners of both password
fields. For the password field
on focusout, it calls the methods
that validates the password and
the confirm password field
because the confirm password method
checks if the both passwords match.
So, whenever the password is modified
and the focus leaves the field,
it's necessary to recheck the
confirmPassword field to ensure
that the two fields still match.
If they don't, the validation
message for confirmPassword
will change accordingly.

Call this method in the
initialize method.