Refactor Email Validation
Move the selection of the
email field to the
constructor so that it
be is accessible within
the class.

Also create a new property
that holds the default
message when user tries to
submit an empty field.

Seperate the resposibilities
so that the methods that
check for empty field and
matching pattern are different.

The logic for the empty field
now checks if the field is
not empty. If it is empty,
it sets a custom validation
message and sets an empty
string as the validation
message if it is not.
This would be used later to
prevent the user from trying to
submit the form with an empty
email field.

The previous email validation
happens on focusout and so
the email field remains valid
until the user focuses out -
the user can thus submit the
form with an empty email field
(without interacting with it.)

In validateEmail() change the
if else blocks to reflect the
new methods that check for
the empty field and the
matching pattern.