Check for empty form fields
Instead of checking for an
empty email field alone,
update the method of
take in multiple arguments
(input fields) and check
and check if they are empty.

In the method, set allValid
to be true. Loop through
the inputFields
and set an isNotEmpty constant
that checks if the input
field is empty. For each
that is empty,
it sets a dafault
custom validity message on it
and changes the value of allValid
to false. Else it sets its
Custom validity message to
an empty string.

It then reports the validity of
the each input field and
returns allNotEmpty.

This would be used on click
of the submit button, to
prevent users from submitting
the form with an empty field.

Update validateEmail with this
new method and call emailInput
as its argument.