Refactor the password fields validation
When the submit button is clicked,
the password and confirmPassword
fields will also have to be
validated.

In the validatePasswordField(),
save the password regular
expression test
to a constant and use the
constant in the if else block.
Return the constant.

In the
validateConfirmPasswordField(),
save the match passwords check
and use the
constant in the if else block.
Return the constant.

With these two methods return
values, the submit button
can call them on click and not
have to rely on the focusout
event of the password fields
for validation.