Validate on submit
Create a validateOnSubmit
method that would
validate all form fields
for all constraints
when the submit button is
clicked.

In the method, attach a click
event listener to the
submit button.

When the submit button is
clicked, checks if all
the inputs are empty -
(calls the
checkForEmptyInput() method
in the IF block
with the necessary form fields
as args) and prevents the
submit button from submitting.

In the if else block, it
checks if all the form fields
meet their respective
constraints and if so, it
also prevents the button from
submitting.

Else (all constraints
are satisfied), it logs a success
message on the console.

Call the validateOnSubmit method
in the initialize method.