Add friendly message on submit
In the index.html, add a span that
will hold a success message.

In the index.css, style the span
with a class that will be added
to the span.

In the index.js, the validateOnSubmit
method, select the span and clear
its content. When the form fields
have passed their constraints,
add a class (that was mentioned
in the previous paragreph)
and a friendly
textContent to the span.